tests: split consensus tests into separate files for each chain type to avoid non deterministic behaviour due to shared global state in global.rs
